bpf: Invoke tracing progs from a fprobe on archs without direct call
Some architectures may not support ftrace direct calls and can not jump
directly to a BPF trampoline. This provides a fallback built on fprobe
(itself built on a ftrace ops and rethook) so that BPF tracing programs
can still be called on these architectures.

The behavior of the entry and exit hooks somewhat mirrors JITted
trampolines but it has a few implementation differences. First of all,
the code is written in C and architecture independent. Arguments are
converted from a ftrace_regs structure to a BPF context instead of being
read from registers directly. And arguments for fexit are saved in the
rethook nodes pool rather than in a stack frame below the monitored
function.

Signed-off-by: Florent Revest <revest@chromium.org>