rethook: fprobe: Use ftrace_regs instead of pt_regs
On architectures or configs without DYNAMIC_FTRACE_WITH_REGS but with
DYNAMIC_FTRACE_WITH_ARGS, a pt_regs can not necessarily be extracted.
Providing a ftrace_regs gives more flexibility to the fprobe and rethook
handlers.

Kprobe and kretprobe still provide pt_regs as APIs so if they are built
on top of fprobe and rethook, they need to reconstruct a pt_regs on
stack out of the ftrace_regs. Some members of that reconstructed
structure can possibly make no sense.

TODO: We should restore the ftrace_regs from the pt_regs after a
kprobe/kretprobe.

... Alternatively, we could also just break the kprobe and kretprobe API
and pass ftrace_regs to callbacks... This will make a bunch of module
writers unhappy but that's life. We probably want to keep BPF kprobe
programs backward compatible though, so we could have the BPF verifier
transparently overwrite accesses to ftrace_regs as accesses to pt_regs.

Signed-off-by: Florent Revest <revest@chromium.org>