fprobe: Allow use without CONFIG_DYNAMIC_FTRACE_WITH_REGS
By registering to ftrace with SAVE_REGS_IF_SUPPORTED we specify that
fprobe wants full regs if they are available but is also happy to
execute if they are not. For example, on architectures with
CONFIG_DYNAMIC_FTRACE_WITH_ARGS or for architectures without anything
saved. In that case, the users of fprobe need to handle the absence of
ftrace_regs.

TODO: Maybe the nicer fix would be to add a FTRACE_OPS_FL_SAVE_ARGS
instead...

Signed-off-by: Florent Revest <revest@chromium.org>