fprobe: Allow entry handlers to skip the execution of exit handlers
In some circumstances, it may not be known at fprobe attachment type
whether both the entry and exit callbacks should be called so both are
registered to fprobe.

However, at the time the entry callback is called, it may know that
the exit callback is superfluous. To avoid the overhead of calling an
empty exit callback, this patch introduces a mean for entry callbacks to
skip the exit callback.

Signed-off-by: Florent Revest <revest@chromium.org>