[WIP]arm64, ftrace: Implement direct calls
In the case when the direct called trampoline is in range, the ftrace
patchsite is just replaced with a call to the trampoline.

If it is out of range, the ftrace patchsite will jump to a "direct
caller" trampoline which will only read the trampoline's address from
the patchsite's literal area. (originally introduced for direct ops)

If an ftrace ops is also called for that function, then the ftrace
trampoline returns back into the direct called trampoline.

We need to be extremely careful with the order in which we modify the
literal and the jump in the call site because of concurency issues. This
patch addresses this by... well, actually it doesn't. ¯\_(ツ)_/¯ Hence
the WIP tag...