ftrace: Remove the legacy direct-call API
There are no more in-tree users of this API. Only the *_multi equivalent
are used nowadays. This is a lot of dead-code to maintain for no reason.

Also, this API relies on a single ops being used for all direct calls
registered through it. However, in a later refactoring patch, it will be
beneficial to have a 1:1 relationship between ops and their direct
called trampolines.

Signed-off-by: Florent Revest <revest@chromium.org>