ftrace: Add a macro to forge an incomplete pt_regs from a ftrace_regs
When built with CONFIG_DYNAMIC_FTRACE_WITH_ARGS, struct ftrace_regs
contains a subset of registers relevant for tracing.

On most architectures, ftrace_regs really just wraps a pt_regs in which
some fields are not set. On arm64, ftrace_regs has a totally different
layout than a pt_regs.

Calling the ftrace_get_regs macro returns NULL to indicate that a
complete pt_regs can not be extracted and avoid that the user shoots
themselves in the foot by accessing fields which are not set.

However, there are situations in which the user may actually want to get
an incomplete pt_regs and deal with the consequences. (eg: to preserve
backward compatibility of certain API boundaries)

To that effect, this patch introduces a macro that lets the user
construct or extract an incomplete pt_regs out of a struct ftrace_regs.

Signed-off-by: Florent Revest <revest@chromium.org>